Use .pigment.tsx as next.js route pages directly
Also add a basic about page.
