--- docs/src/app/layout.tsx
@@ -1,84 +1,0 @@
-import '@pigment-css/react-new/styles.css';
-import './globals.css';
-import './syntax.css';
-
-import * as React from 'react';
-import { Metadata, Viewport } from 'next';
-import { Inter } from 'next/font/google';
-
-import { bodyCls, htmlCls } from './layout.pigment';
-
-const inter = Inter({
-  variable: '--font-inter',
-  subsets: ['latin'],
-});
-
-export default async function Layout({ children }: React.PropsWithChildren) {
-  return (
-    // suppressHydrationWarning is needed because we immediately modify the html on client
-    // before React hydration happens
-    <html lang="en" className={`${htmlCls}`} data-theme="system" suppressHydrationWarning>
-      <head>
-        <script
-          id="theme-selector"
-          dangerouslySetInnerHTML={{
-            __html: `const mode = window.localStorage.getItem('mode');
-if (mode === 'system' || mode === 'dark') {
-  document.documentElement.dataset.theme = mode;
-}`,
-          }}
-        />
-      </head>
-      <body className={`${inter.variable} ${bodyCls}`}>{children}</body>
-    </html>
-  );
-}
-
-export const metadata: Metadata = {
-  title: {
-    template: `%s · ${process.env.APP_NAME}`,
-    default: process.env.APP_NAME as string,
-  },
-  twitter: {
-    site: '@PigmentCSS',
-    card: 'summary_large_image',
-  },
-  openGraph: {
-    type: 'website',
-    locale: 'en_US',
-    title: {
-      template: `%s · ${process.env.APP_NAME}`,
-      default: process.env.APP_NAME as string,
-    },
-    ttl: 604800,
-  },
-  applicationName: process.env.APP_NAME,
-  icons: {
-    icon: [
-      {
-        url:
-          process.env.NODE_ENV === 'production' ? '/static/favicon.ico' : '/static/favicon-dev.ico',
-        sizes: '32x32',
-      },
-      {
-        url:
-          process.env.NODE_ENV === 'production' ? '/static/favicon.svg' : '/static/favicon-dev.svg',
-        sizes: 'any',
-        type: 'image/svg+xml',
-      },
-    ],
-    shortcut: {
-      url: '/static/apple-touch-icon.png',
-      sizes: '180x180',
-    },
-    apple: {
-      url: '/static/apple-touch-icon.png',
-      sizes: '180x180',
-    },
-  },
-};
-
-export const viewport: Viewport = {
-  initialScale: 1,
-  width: 'device-width',
-};
